pyproject.toml
[tool.poetry]
name = "finalsurge2intervals"
version = "0.0.1"
description = ""
authors = ["Evgeny <iam@freekode.org>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
python-dateutil = "^2.8.2"
PyYAML = "^6.0"
intervalsicu = "^0.1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
